set correct swift dispersion tenant
In kilo, the swift dispersion user is created as a member of
the 'service' tenant.  However, we are populating the
swift-dispersion.conf as though it was a member of the admin tenant (as
it was in juno).  This commit fixes that, and makes
swift-dispersion-report (and any monitoring plugins that rely on it)
work again

Change-Id: I0e9b9ccaa487ee0242346ee279821aa09dd5b022
Closes-Bug: #1479006