Fix inconsistent gate and lint failures for os-swift
Gate was failing due to a race condition whereby the partition is not ready
and so the first mounted partitions fails to mount - causing hosts to fail.

Adding a retry/delay on failures in the mount task should resolve this issue.

ansible-lint failures were occurring due to a recent release of ansible-lint
Since the gate failure cannot be merged without the linting fix, this change
also contains I0cc973867d9c279eb6ea05decc5ebccfaa19c7b0 to break the co-
dependency between this and the other patch.

Change-Id: I3f4daaed764c4cad8a07408fedf8d0cba570a344