Enable pypy to be used with swift
The intention of this change is to enable an option to increase swift
performance without changing any of the swift internals. This change is
in-line with documentation published by the good folks at swiftstack and
intel. This change will only effect swift when its installed in a
virtual environment which is a self-imposed limitation such that we're
not messing with the system's default Python.

New Options:
  - A new option `swift_pypy_enabled` has been added to enable or
    disable the pypy interpreter for swift. The default is "false".
  - A new option `swift_pypy_archive` has been added to allow a
    pre-built pypy archive to be downloaded and moved into place. This
    option is a dictionary and contains the URL and SHA256 as keys.

Reference: [ https://software.intel.com/en-us/blogs/2016/05/06/doubling-the-performance-of-openstack-swift-with-no-code-changes ].

Change-Id: I045e8b97d29b96e2c2450761efa1c2668de12a75
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>